strsep rewrite. warnings eliminated

+++ tnt/tnt_auth.c
@@ -79,7 +79,7 @@ tnt_auth_raw(struct tnt_stream *s, const char *user, int ulen,
 	char salt[64], scramble[TNT_SCRAMBLE_SIZE];
 	if (!guest) {
 		data = mp_encode_array(data, 2);
-		data = mp_encode_str(data, "chap-sha1", strlen("chap-sha1"));
+		data = mp_encode_str(data, "chap-sha1", (uint32_t)strlen("chap-sha1"));
 		data = mp_encode_strl(data, TNT_SCRAMBLE_SIZE);
 		base64_decode(base64_salt, TNT_SALT_SIZE, salt, 64);
 		tnt_scramble_prepare(scramble, salt, pass, plen);
