+++ tnt/tnt_execute.c
@@ -62,7 +62,7 @@ tnt_execute_int(struct tnt_stream *s, const char *expr, size_t elen, struct tnt_
 
 	data = mp_encode_map(data, 2);
 	data = mp_encode_uint(data, TNT_SQL_TEXT);
-	data = mp_encode_strl(data, elen);
+	data = mp_encode_strl(data, (uint32_t)elen);
 	v[2].iov_base = body_start;
 	v[2].iov_len = data - body_start;
 	v[3].iov_base = (void *) expr;
